The fourth revision (#6)

+++ core/simd_ops.h
@@ -1,12 +1,14 @@
 #ifndef SIMD_OPS_H
 #define SIMD_OPS_H
 
+#include <stddef.h>
+
 #ifdef __cplusplus
 extern "C" {
 #endif
 
-// Normalizes a single vector using AVX instructions
-void avx_normalize(float *vec, int len);
+// Normalizes a single vector using AVX.
+void avx_normalize(float *vec, size_t len);
 
 #ifdef __cplusplus
 }
